requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
distlib==0.3.4
Django==3.2.11
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.4.2
idna==3.3
lark-parser==0.12.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.21.5
pandas==1.3.5
platformdirs==2.4.1
protobuf==3.19.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
required==0.4.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
